Share search and sort with My Library
Still kind of hacky, relies on hardcoded ids.

TODO try improving with EventBus?
